Add more structure for link collecting and checks
and move checker functions to own submodules
