+++ libimagentrymarkup/src/asciidoc/check.rs
@@ -1,0 +1,6 @@
+use libimagstore::store::Entry;
+
+pub fn is_asciidoc(e: &Entry) -> bool {
+    false
+}
+
