+++ libimagentrymarkup/src/bbcode/check.rs
@@ -1,0 +1,6 @@
+use libimagstore::store::Entry;
+
+pub fn is_bbcode(e: &Entry) -> bool {
+    false
+}
+
