+++ libimagentrymarkup/src/latex/check.rs
@@ -1,0 +1,6 @@
+use libimagstore::store::Entry;
+
+pub fn is_latex(e: &Entry) -> bool {
+    false
+}
+
