+++ libimagentrymarkup/src/textile/check.rs
@@ -1,0 +1,6 @@
+use libimagstore::store::Entry;
+
+pub fn is_textile(e: &Entry) -> bool {
+    false
+}
+
